pyproject.toml
[tool.poetry]
name = "Piranhabon"
version = "0.1.0"
description = "convert my NAIST timetable into Google Calendar"
authors = ["5ebec <hash.a5ebec@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.24.0"
beautifulsoup4 = "^4.9.3"
google-api-python-client = "^1.12.3"
google-auth-httplib2 = "^0.0.4"
google-auth-oauthlib = "^0.4.1"
python-dotenv = "^0.14.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
